Add b2aExtraLife to Complete Systems
This is a smaller system I created to use for my Extra Life 2016 stream.
It will connect to the Extra Life JSON to keep track of donation totals
and get a list of donors. The bundle can also keep a list of the games
you are playing to display in the lower third as well as any other
messages you want to show. I think it could be a nice addition to the
list as lots of streamers over on the Extra Life forums are always
looking for overlays to track this data. Thanks for your consideration!

https://github.com/cshomo11/b2aExtraLife